requirements.txt
streamlit
python-dotenv
gspread
oauth2client
fastapi
uvicorn
pydantic
requests
groq
# ai-chatbot-app/requirements.txt  (append to the end)
#psycopg2-binary==2.9.9
beautifulsoup4==4.12.3
requests==2.32.3
supabase>=2.4.0
langchain-community>=0.3.0
langchain-core>=0.3.0
langchain-text-splitters>=0.3.0
langchain-groq>=0.2.0
sentence-transformers>=3.0.0
selenium>=4.21.0
websockets>=12.0
realtime>=1.0.0
pyarabic
langdetect